# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Regina caeli API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/patient/all-sessions": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "operationId": "7d5f9559c8003ca88301bd9d0dd5b595",
                "responses": {
                    "200": {
                        "description": "Booking successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Code Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Access",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Booking already exists",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/patient/add-a-session": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "operationId": "502d9a7996aa3e2205330d074f8decfc",
                "parameters": [
                    {
                        "name": "booking_start",
                        "in": "query",
                        "description": "booking_start",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "service_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "booking_for_self",
                        "in": "query",
                        "description": "booking_for_self",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "booked_by_id",
                        "in": "query",
                        "description": "booked_by_id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Code Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized Access",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Booking already exists",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}